% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{discover_process}
\alias{discover_process}
\alias{discover_process.eventlog}
\title{Discover process}
\usage{
discover_process(log)

\method{discover_process}{eventlog}(log)
}
\arguments{
\item{log}{Event log}
}
\description{
Discover process
}
\section{Methods (by class)}{
\itemize{
\item \code{discover_process(eventlog)}: eventlog

}}
