% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_petrinet.R
\name{to_petrinet}
\alias{to_petrinet}
\title{Convert BPMN model to petrinet notation}
\usage{
to_petrinet(bpmn)
}
\arguments{
\item{bpmn}{A bpmn model, as discovered using \code{discover()}}
}
\value{
A marked petrinet, as defined by petrinetR
}
\description{
Convert BPMN model to petrinet notation
}
