% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{discover}
\alias{discover}
\alias{discover.eventlog}
\title{Discover process}
\usage{
discover(log, loop_strategy)

\method{discover}{eventlog}(log, loop_strategy = "skip")
}
\arguments{
\item{log}{Event log}
}
\description{
Discover process
}
\section{Methods (by class)}{
\itemize{
\item \code{discover(eventlog)}: eventlog

}}
